feat(auth): extract rise-backend-auth pure-core crate (Phase 0)
Centralize all Rise token signing, verification, and matching into a new
pure-core workspace crate (crates/rise-backend-auth) — no reqwest/axum/sqlx/
tokio/anyhow/regex. This establishes the single home for auth-token logic and
the two mandated entry points:

- verify_rise_jwt(token) -> RiseToken{Session,Ingress}: the one path that
  verifies a Rise-issued JWT, disambiguating by algorithm. The legacy
  verify_user_jwt / verify_jwt_skip_aud are thin adapters over it.
- verify_external_jwt(token, issuer, &dyn JwksKeySource) -> ExternalClaims: the
  one path that validates an arbitrary external JWT. JWKS fetching is abstracted
  behind the JwksKeySource trait; rise-deploy's JwtValidator implements it
  (reqwest + SSRF + cache stay in rise-deploy).

The crate also owns RiseClaims/WorkloadClaims/ExternalClaims, the RiseTokenSigner,
the pure matchers (validate_custom_claims, glob, ControllerIdentity +
match_controller_identity/build_controller_indexes/validate_controller_id), and
is_rise_issued_jwt.

No externally observable behavior change. Two forced, behavior-preserving deltas
required by the pure-core boundary: the groups DB fetch is hoisted out of
sign_user_jwt/sign_ingress_jwt into the 4 caller sites (these signers are now
synchronous and take groups: Option<Vec<String>>), and the controller matchers
return the crate's AuthError instead of anyhow (match_controller_identity is
unchanged). validate_controller_id is hand-rolled (no regex) reproducing the
original CONTROLLER_ID_RE rules, locked by ported tests.

The crate is backend-only (optional, backend-gated, strip-on-publish), mirroring
the existing resource crates. AccessClaims and the exchange endpoint are NOT
introduced here (Phase 1).
