+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     ".",
+    "crates/rise-backend-auth",
     "crates/rise-resource-api",
     "crates/rise-resource-store",
 ]
@@ -9,11 +10,15 @@ resolver = "2"
 [workspace.dependencies]
 anyhow = "1.0.100"
 async-trait = "0.1"
+base64 = "0.22"
 chrono = { version = "0.4.38", features = ["serde"] }
+jsonwebtoken = { version = "10.0", features = ["rust_crypto"] }
 rand = "0.10.0"
+rsa = "0.9"
 schemars = { version = "1", features = ["chrono04", "uuid1"] }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
+sha2 = "0.11"
 sqlx = { version = "0.8", features = ["runtime-tokio-native-tls", "postgres", "uuid", "chrono", "migrate", "json"] }
 thiserror = "2.0"
 tokio = { version = "1.48.0", features = ["full"] }
@@ -62,6 +67,7 @@ cli = [
 ]
 backend = [
     # Server core (former "server" feature)
+    "dep:rise-backend-auth", # strip-on-publish
     "dep:rise-resource-api", # strip-on-publish
     "dep:rise-resource-store", # strip-on-publish
     "dep:config",
@@ -103,6 +109,7 @@ backend = [
 [dependencies]
 # --- BEGIN BACKEND (stripped before `cargo publish` in CI; see .github/workflows/ci.yml) ---
 # Workspace crates (backend only)
+rise-backend-auth = { path = "crates/rise-backend-auth", optional = true }
 rise-resource-api = { path = "crates/rise-resource-api", optional = true }
 rise-resource-store = { path = "crates/rise-resource-store", optional = true }
 # --- END BACKEND ---
@@ -110,7 +117,7 @@ rise-resource-store = { path = "crates/rise-resource-store", optional = true }
 # Common (always enabled - shared across features)
 anyhow = "1.0.100"
 async-stream = "0.3"
-base64 = "0.22"
+base64 = { workspace = true }
 bytes = "1.9"
 chrono = { workspace = true }
 futures = "0.3"
@@ -126,8 +133,8 @@ uuid = { workspace = true }
 # Shared: CLI + Server
 axum = "0.8.7"
 axum-extra = { version = "0.10", features = ["query"] }
-jsonwebtoken = { version = "10.0", features = ["rust_crypto"] }
-sha2 = "0.11"
+jsonwebtoken = { workspace = true }
+sha2 = { workspace = true }
 urlencoding = "2.1"
 
 # CLI-only dependencies
@@ -147,7 +154,7 @@ sqlx = { version = "0.8", features = ["runtime-tokio-native-tls", "postgres", "u
 tower = { version = "0.5.2", optional = true }
 tower-http = { version = "0.6.7", features = ["trace", "cors"], optional = true }
 regex = { version = "1.11", optional = true }
-rsa = { version = "0.9", optional = true }
+rsa = { workspace = true, optional = true }
 
 # Server: Encryption (non-KMS mode)
 aes-gcm = { version = "0.10", optional = true }
