+++ src/server/auth/middleware.rs
@@ -12,26 +12,7 @@ use crate::db::{service_accounts, users, User};
 use crate::server::auth::context::VerifiedExternalToken;
 use crate::server::auth::cookie_helpers;
 use crate::server::state::AppState;
-
-/// Check if a JWT issuer is a Rise-issued JWT
-///
-/// Rise JWTs have `iss` set to the Rise public URL (e.g., "https://rise.example.com").
-/// This helper checks for exact match or scheme prefix match.
-fn is_rise_issued_jwt(issuer: &str, public_url: &str) -> bool {
-    // Exact match
-    if issuer == public_url {
-        return true;
-    }
-
-    // Check if issuer starts with the public_url's base (handles port differences)
-    if let Some(public_base) = public_url.strip_suffix(|c: char| c.is_ascii_digit() || c == ':') {
-        if issuer.starts_with(public_base) {
-            return true;
-        }
-    }
-
-    false
-}
+use rise_backend_auth::is_rise_issued_jwt;
 
 /// Extract Bearer token from Authorization header
 pub(crate) fn extract_bearer_token(headers: &HeaderMap) -> Option<String> {
